Update package version and improve documentation
This commit updates the version of the package from '1.0.3' to '1.0.4' and makes several improvements to the documentation in the readme file.

The readme changes include clearer examples of how to use the Handlebars library and update installation instructions. The description of the library has been rephrased to emphasize its focus on backend generation. A detailed example has also been added to demonstrate the available features of the library.

Moreover, two binary files, 'small.jpg' and 'social.jpg', have been updated, but the specific changes to these files are not detailed in the diff.